+++ src/pages/PlaygroundPage.svelte
@@ -5,118 +5,19 @@
  -->
 
 <script lang="ts">
-  import Gesture from '../script/domain/Gesture';
-  import Model from '../script/domain/Model';
-  import LayersModelTrainer from '../script/mlmodels/LayersModelTrainer';
-  import {
-    classifier,
-    engine,
-    gestures,
-    liveAccelerometerData,
-  } from '../script/stores/Stores';
-  import AccelerometerClassifierInput from '../script/mlmodels/AccelerometerClassifierInput';
-  import PlaygroundGestureView from '../components/playground/PlaygroundGestureView.svelte';
-  import playgroundContext from '../components/playground/PlaygroundContext';
   import PlaygroundLog from '../components/playground/PlaygroundLog.svelte';
   import MicrobitAccelerometerDataSynthesizer from '../components/playground/inputSynthesizer/MicrobitAccelerometerDataSynthesizer.svelte';
   import LiveDataBufferUtilizationPercentage from '../components/playground/LiveDataBufferUtilizationPercentage.svelte';
-  import StaticConfiguration from '../StaticConfiguration';
-
-  const getRandomGesture = (): Gesture => {
-    return gestures.getGestures()[
-      Math.floor(Math.random() * gestures.getNumberOfGestures())
-    ];
-  };
-
-  const model: Model = classifier.getModel();
-  const trainModelButtonClicked = () => {
-    playgroundContext.addMessage('training model...');
-    model
-      .train(new LayersModelTrainer(StaticConfiguration.layersModelTrainingSettings))
-      .then(() => {
-        playgroundContext.addMessage('Finished training!');
-      });
-  };
-
-  const predictButtonClicked = () => {
-    const randGesture = getRandomGesture();
-    playgroundContext.addMessage(
-      'Predicting on random recording of: ' + randGesture.getName(),
-    );
-    const xs = randGesture.getRecordings()[0].data.x;
-    const ys = randGesture.getRecordings()[0].data.y;
-    const zs = randGesture.getRecordings()[0].data.z;
-    const input = new AccelerometerClassifierInput(xs, ys, zs);
-    classifier.classify(input).then(() => {
-      playgroundContext.addMessage('Finished predicting');
-    });
-  };
+  import StoresDisplay from '../components/playground/StoresDisplay.svelte';
+  import EngineInteractionButtons from '../components/playground/EngineInteractionButtons.svelte';
 </script>
 
 <div class="flex flex-col p-5">
   <div class="flex">
-    <div class="flex flex-col flex-shrink">
-      <div>
-        <p class="text-2xl mt-2">Model store</p>
-        <p class="whitespace-pre">
-          {JSON.stringify($model, null, 2).substring(
-            2,
-            JSON.stringify($model, null, 2).length - 1,
-          )}
-        </p>
-      </div>
-      <div>
-        <p class="text-2xl mt-2">Classifier store</p>
-        <p class="whitespace-pre">
-          {JSON.stringify($classifier, null, 2).substring(
-            2,
-            JSON.stringify($classifier, null, 2).length - 1,
-          )}
-        </p>
-      </div>
-      <div>
-        <p class="text-2xl mt-2">Engine store</p>
-        <p class="whitespace-pre">
-          {JSON.stringify($engine, null, 2).substring(
-            2,
-            JSON.stringify($engine, null, 2).length - 1,
-          )}
-        </p>
-      </div>
-      <div>
-        <p class="text-2xl mt-2">LiveData store</p>
-        <p class="whitespace-pre">
-          {JSON.stringify($liveAccelerometerData, null, 2).substring(
-            2,
-            JSON.stringify($liveAccelerometerData, null, 2).length - 1,
-          )}
-        </p>
-      </div>
-      <div>
-        <p class="text-2xl mt-2">Gestures (without recordings/output)</p>
-        {#each $gestures as { ID }}
-          <PlaygroundGestureView gesture={gestures.getGesture(ID)} />
-        {/each}
-      </div>
-    </div>
+    <StoresDisplay />
     <div class="flex-grow" />
     <div class="flex flex-col w-100">
-      <button class="border-1 p-2 m-1" on:click={trainModelButtonClicked}
-        >train model!</button>
-      <button class="border-1 p-2 m-1" on:click={predictButtonClicked}
-        >Predict random gesture!</button>
-      <button
-        class="border-1 p-2 m-1"
-        on:click={() => {
-          playgroundContext.addMessage('Starting engine');
-          engine.start();
-        }}>Start engine!</button>
-      <button
-        class="border-1 p-2 m-1"
-        on:click={() => {
-          playgroundContext.addMessage('Stopping engine');
-          engine.stop();
-        }}>Stop engine!</button>
+      <EngineInteractionButtons />
       <PlaygroundLog />
     </div>
   </div>
